% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_simultanee.R
\name{bisig_plot}
\alias{bisig_plot}
\title{bisig_plot draws the plots of the bivaraite signal on the same plot (scale free)}
\usage{
bisig_plot(x, rupt = NULL, mu = NULL, pop = NULL)
}
\arguments{
\item{x}{the signal to be plotted}
\item{mu}{the mu}
\item{pop}{the pop}

\item{rupt}{optionnal, if given add vertical lines at change points (rupt should)}
}
\value{
no value
}
\description{
bisig_plot draws the plots of the bivaraite signal on the same plot (scale free)
}

