% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EM.algo_simultanee.R
\name{EM.algo_simultanee}
\alias{EM.algo_simultanee}
\title{EM.algo_simultanee caculates the MLE of phi for given change-point instants}
\usage{
EM.algo_simultanee(x, rupt, P, phi)
}
\arguments{
\item{rupt}{ahe sequence of change points}

\item{P}{number of clusters}

\item{phi}{starting value for the  parameter}
}
\value{
a list with  phi, the MLE, tau =(taukj) the probability for segment k to belong to classe,lvinc = lvinc,empty = empty,dv = dv
}
\description{
EM.algo_simultanee caculates the MLE of phi for given change-point instants
}

