% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EM.init_simultanee.R
\name{EM.init_simultanee}
\alias{EM.init_simultanee}
\title{EM.init_simultanee proposes an inial value for the EM algorithm  based on a hierarchical clustering  algorithm (ascending)}
\usage{
EM.init_simultanee(x, rupt, K, P)
}
\arguments{
\item{x}{the bivariate signal}

\item{rupt}{the  change point instants, data.frame}

\item{K}{an integer value indicating the number of segments}

\item{P}{the number of classes}

}
\description{
EM.init_simultanee proposes an inial value for the EM algorithm  based on a hierarchical clustering  algorithm (ascending)
}

