% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Mstep_simultanee.R
\name{Mstep_simultanee}
\alias{Mstep_simultanee}
\title{Mstep_simultanee computes the MLE within the EM framework}
\usage{
Mstep_simultanee(x, rupt, tau, phi)
}
\arguments{
\item{x}{the bivariate signal}

\item{rupt}{the rupture dataframe}

\item{tau}{the K*P matrix containing posterior probabilities of membership to clusters}

\item{phi}{the parameters of the mixture}
}
\value{
phi the updated value of the pramaeters
}
\description{
Mstep_simultanee computes the MLE within the EM framework
}

