% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{neighbors}
\alias{neighbors}
\title{neighbors tests whether neighbors of point k,P can  be used to re-initialize the EM algorithm and to improve the log-likelihood.}
\usage{
neighbors(x, L, k, param, ibp = c(), P, lmin, sameSigma = TRUE)
}
\arguments{
\item{x}{the initial dataset}

\item{L}{the likelihood}

\item{k}{the points of interest}

\item{ibp}{to deal with forbidden path, not currently implemented}

\item{P}{the number of class}

\item{lmin}{minimal size of the segment to be implemented}
}
\value{
smoothin likelihood
}
\description{
neighbors tests whether neighbors of point k,P can  be used to re-initialize the EM algorithm and to improve the log-likelihood.
}

