\name{picard}
\alias{picard}
\alias{picard.default}
\alias{picard.ltraj}
\alias{plot.picard}
\alias{print.picard}
\title{
  Simultaneous Segmentation -- Clustering of Signals using the method of
  Picard (2006)
}
\description{
  These functions allow to perform a segmentation and clustering of a
  time series using the hybrid algorithm of Picard (2006).  The function
  \code{picard} implements this approach either from a series of
  observation or from an animal trajectory.  The function \code{plot}
  can then be used to plot the results.
}
\usage{
picard(x, \dots)
\method{picard}{default}(x, P, Kmax, pos=NULL, \dots)

\method{picard}{ltraj}(x, P, Kmax, which=c("x","y"), \dots)

\method{print}{picard}(x, \dots)

\method{plot}{picard}(x, axes = FALSE, number = NULL,
                      addparam = FALSE, bg = TRUE, \dots)

}
\arguments{
  \item{x}{for \code{picard.default}, a matrix with two columns containing the
    successive observations building up the bivariate series (the two
    columns are the two variables).  For \code{picard.ltraj}, an object
    of class \code{ltraj}.  For \code{plot}, an object of class
    \code{"picard"}.}
  \item{P}{an integer value indicating the number of classes used for
    the segmentation.}
  \item{Kmax}{an integer value indicating the maximum number of segments
    expected in the series.}
  \item{pos}{optionnally, a vector containing the same number of
    elements as x, corresponding to the "coordinates" of each
    observation in the series (e.g. the dates corresponding to each
    observation).}
  \item{which}{a character vector of length 2 indicating which variables
    of the object ltraj are to be used in the segmentation process
    (these variables may be selected in the infolocs component of the
    object).}
  \item{axes}{a logical value indicating whether the x and y axes are to
    be plotted.}
  \item{number}{optionally, an integer value indicating which of the two
    variable is to be plotted (by default, both variables are stored).}
  \item{addparam}{a logical value indicating whether the parameters
    (mean and sd) should be plotted.}
  \item{bg}{a logical value indicating whether colored rectangle
    indicating the model corresponding to each segment should be drawn
    in the background}
  \item{\dots}{additional arguments to be passed from or to other
    functions}
}
\details{
  The method of Picard (2006)... to be continued.
}
\value{
  The function \code{picard} returns a list containing the following
  elements:
  \item{Linc}{a vector containing the log-likelihood for the best
    segmentation of the signal with K-segment}
  \item{param}{a list containing the parameters of the segmentation with
    one element per value of K; each element is itself a list
    containing: (i) a list named \code{phi} containing the parameters of
    the models, i.e. with a component named \code{mu} containing the
    mean of each variable (column) for each model (row), a component
    named \code{sigma} containing the standard deviation of each
    variable (column) for each model (row) and a component named
    \code{prop} containing the proportion of observations generated by
    each model; (ii) a matrix named \code{rupt} containing the indices
    of the limits (start, end) of the K segments (one per row); (iii) a
    matrix named \code{tau} containing the likelihood of each model
    (row) for each segment (column)}
}
\references{
  Picard, F. 2006. Process segmentation/clustering. Application to the
  analysis of CGH microarray data. PhD thesis, University Paris XI,
  Orsay.
}
\author{
  Clement Calenge \email{clement.calenge@oncfs.gouv.fr}
  Marie-Pierre Etienne \email{marie.etienne@agroparistech.fr}
  Emilie Lebarbier \email{emilie.lebarbier@agroparistech.fr}
}
\examples{

############################################
##
## Example 1:

set.seed(879)

## Generates a series with 3 segments
x1 <- c(rnorm(30, 3, 1), rnorm(20, 8, 1), rnorm(10, 3, 1))
x2 <- c(rnorm(30, -2, 2), rnorm(20, 5, 2), rnorm(10, -2, 2))
xk <- cbind(x1,x2)

## fits the segmentation:
seg <- picard(xk, 2, Kmax = 5)

## Plot the result:

## the models
plot(seg)

## show the parameters for the first variable
plot(seg, number = 1, bg = FALSE, addparam = TRUE)

## show the parameters for the first variable
plot(seg, number = 2, bg = FALSE, addparam = TRUE)

############################################
##
## Example 2: no clear sequence

set.seed(980)
x <- cbind(rnorm(50), rnorm(50))

seg2 <- picard(x, 3, 8)

## Les segments sont definis par deux obs a chaque fois
## bizarre, non?
plot(seg2)

## Au passage
## Les resultats sont identiques quand on utilise
## hybrid_simultanee:
oo <- hybrid_simultanee(t(x), 3, 8)

## Je retrouve la meme chose avec des vrais trajets

############################################
##
## Example 3: wild boar

data(puechcirc)
pu <- puechcirc[1]

seg3 <- picard(pu, 2, 7)

plot(seg3)

## bizarre, non?

}
\keyword{cluster}
\keyword{chron}
