\name{segTraj-package}
\alias{segTraj-package}
\alias{segTraj}
\docType{package}
\title{
  Segmentation of Bivariate Time Series
}
\description{
  Segmentation and clustering of a bivariate time series.
}
\details{
\tabular{ll}{
Package: \tab segTraj\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-20\cr
License: \tab What license is it under?\cr
}
}
\author{
Marie-Pierre Etienne, Emilie Lebarbier, Clement Calenge and Franck Picard

Maintainer: Marie-Pierre Etienne <marie.etienne@agroparistech.fr>
}
\references{
  Picard, F. 2006. Process segmentation/clustering. Application to the
  analysis of CGH microarray data. PhD thesis, University Paris XI,
  Orsay.
}
\keyword{ package }
\examples{
## No examples
}
